Port int64/uint64 to int/uint for 32-bit compatibility
This commit ports the internal use of fixed-width 64-bit integers
(int64, uint64) to machine-sized integers (int, uint) across the
entire codebase. This ensures that mgmt is buildable on 32-bit
architectures while still respecting external library requirements
for 64-bit types.

Key changes:
- Updated lang/types.Value interface and IntValue struct to use int.
- Updated engine.MetaParams.Delay to use uint.
- Updated various resource fields (timeouts, ports, etc.) to use uint/int.
- Modernized string conversions to use architecture-aware strconv functions.
- Ensured explicit casts to int64 for external library calls (e.g., time.Unix).
- Updated the test suite to reflect the new integer types.

+++ lang/core/convert/to_int_test.go
@@ -36,7 +36,7 @@ import (
 	"github.com/purpleidea/mgmt/lang/types"
 )
 
-func testToInt(t *testing.T, input float64, expected int64) {
+func testToInt(t *testing.T, input float64, expected int) {
 
 	got, err := ToInt(context.Background(), []types.Value{&types.FloatValue{V: input}})
 	if err != nil {
